\name{CmmD_from_community_structures()}
\alias{CmmD_from_community_structures}
\title{
CmmD_from_community_structures
}
\description{
Compute CmmD analysis for network community structure over precomputed MolTi community detection algorithm outputs (Didier et al. 2018)
}
\usage{
CmmD_from_community_structures(nodelist=NULL,community_structures,resolution_start,resolution_end,interval)
}
\arguments{
  \item{nodelist}{Optional. If given, a list with the unique nodes that we want to appear in the final output. If not given, all nodes of the multiplex will be in the final output (nodelist= NULL)}
  \item{community_structures}{A vector with the paths where the output community structures from MolTi are saved. Easy example: c("Molti_Output/0.5.csv","Molti_Output/1.csv")}
  \item{resolution_start}{Numeric, a real number determining the first gamma resolution parameter used to generate the different MolTi's outputs}
  \item{resolution_end}{Numeric, a real number determining the last gamma resolution parameter used to generate the different MolTi's outputs}
  \item{interval}{Numeric, a real number determining the interval of the resolution parameter used to generate the different MolTi's outputs}
}
\details{
Compute CmmD analysis for a set of given community structures generated with MolTi (or have MolTi's format) (Didier et al. 2018).
}
\value{
  \item{gene_community_matrix}{A matrix where the rows correspond to the unique nodes of all layers and the colums identify the resolution parameter used in each analysis. The values in the matrix describe the ID of the community where the node is found in each analysis. Each of the columns have its own way of naming the communities found, as a result of the difference in the number of communities found. Therefore, being at community "2" at resolution parameter 2.5 and being at community "2" at resolution parameter 5 does not mean that the node is kept within the same community during the analysis.}
  \item{l_constant}{A list of vectors with the nodes that are consistently found in the same community through all the analysis. In each vector of the list, you will have nodes that are always found in the same community in all  analyzed values of resolution parameter. The name of the vector describes the communities where the nodes are found. The order of the IDs is the same as in colnames(output$gene_community_matrix).}
  \item{hamming_distance_matrix}{A matrix of dimensions N(nodes) x N(nodes). The hamming distance values between all nodes of the multiplex network. This hamming distance represent the number of analysis where the pair nodes ARE NOT FOUND within the same community.}
}
\references{
Didier, G., Valdeolivas, A., Baudot, A., 2018. Identifying communities from multiplex biological networks by randomized optimization of modularity. F1000Res 7. https://doi.org/10.12688/f1000research.15486.2
}
\author{
Iker Nuñez-Carpintero <iker.nunez@bsc.es>
}
\examples{
#### Example of use:
#### 1. Create a vector with the paths where the community structures are located:
structures <- c("Molti_Output/0.5.csv","Molti_Output/1.csv")

#### 2. In order to analyze the range of resolution parameter 0 to 30, at intervals of 0.5 (e.g: c(0.5,1,1.5,2,2.5,3...,30)):
CmmD_from_community_structures(nodelist=NULL, community_structures=estructuras,resolution_start=0.5,resolution_end=4, interval=0.5)

### Note: The function already gives different names identifying the values of the resolution parameter to each of the community structures generated by MolTi, we highly recommend not to give a prefix to this files apart from the path where you want this files to be saved.

#### OPTIONAL: If you have a specific set of nodes to analyze, you can create a vector with the node names and give it in the argument 'NODELIST':
nodes_to_analyze <- c("A","B","C","D")
CmmD_from_community_structures(nodelist=nodes_to_analyze, community_structures=structures,resolution_start=0.5,resolution_end=4, interval=0.5)
}
